Refactor: Move getUuidByString to separate module
Extracted the UUID generation function into its own utility module (uuidUtils.ts) as requested in PR feedback. This improves code organization and makes the function reusable across the codebase.
